[PP][OPS-11181] back button spec
